课程类别查询查询service

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/xuecheng/content/model/po/CourseMarket.java
@@ -31,7 +31,7 @@ public class CourseMarket implements Serializable {
     /**
      * 现价
      */
-    private Float price;
+    private Integer price;
 
     /**
      * 原价
